Separate types, morphs, and constants by module
